+++ presentation/src/main/java/com/ezlo/mydevices/presentation/home/models/DeviceUiModel.kt
@@ -2,5 +2,5 @@ package com.ezlo.mydevices.presentation.home.models
 
 data class DeviceUiModel(
     val title: String,
-    val deviceSn: Int,
+    val deviceSn: Long,
 )
